+++ flutter_app_project/lib/pages/settings_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../utils/app_logger.dart';
+import '../utils/config_logger.dart';
 
 class SettingsPage extends StatelessWidget {
   const SettingsPage({super.key});
